Move common saga, reducer, actions to App container
