+++ app/containers/App/constants.js
@@ -1,1 +1,4 @@
 export const AD_ID_REGEXP = /.*id=([0-9]*)&access_token.*/;
+export const LOAD_COUNTS = 'Ads/App/LOAD_COUNTS';
+export const LOAD_COUNTS_SUCCESS = 'Ads/App/LOAD_COUNTS_SUCCESS';
+export const LOAD_COUNTS_ERROR = 'Ads/App/LOAD_COUNTS_ERROR';
