+++ app/containers/App/reducer.js
@@ -13,12 +13,21 @@ import {
   LOAD_ADS,
   LOAD_ADS_ERROR,
 } from '../AdsPage/constants';
+import {
+  LOAD_COUNTS,
+  LOAD_COUNTS_SUCCESS,
+  LOAD_COUNTS_ERROR,
+} from './constants';
 
 // The initial state of the App
 export const initialState = {
   loading: false,
   error: false,
   ads: null,
+  counts: {
+    adsCount: null,
+    annotationsCount: null,
+  },
 };
 
 /* eslint-disable default-case, no-param-reassign */
@@ -39,6 +48,24 @@ const appReducer = (state = initialState, action) =>
       case LOAD_ADS_ERROR:
         draft.error = action.error;
         draft.loading = false;
+        break;
+
+      case LOAD_COUNTS:
+        draft.loadingCounts = true;
+        draft.errorCounts = false;
+        break;
+
+      case LOAD_COUNTS_SUCCESS:
+        draft.loadingCounts = false;
+        draft.counts = {
+          adsCount: action.adsCount,
+          annotationsCount: action.annotationsCount,
+        };
+        break;
+
+      case LOAD_COUNTS_ERROR:
+        draft.errorCounts = action.error;
+        draft.loadingCounts = false;
         break;
     }
   });
